Move `Enhancements` parser / `Rule` into separate files
