+++ src/sentry/grouping/enhancer/__init__.py
@@ -12,9 +12,6 @@ import msgpack
 import sentry_sdk
 import zstandard
 from django.core.cache import cache
-from parsimonious.exceptions import ParseError
-from parsimonious.grammar import Grammar
-from parsimonious.nodes import NodeVisitor
 from sentry_ophio.enhancers import Cache as RustCache
 from sentry_ophio.enhancers import Component as RustComponent
 from sentry_ophio.enhancers import Enhancements as RustEnhancements
@@ -26,18 +23,10 @@ from sentry.stacktraces.functions import set_in_app
 from sentry.utils import metrics
 from sentry.utils.hashlib import hash_value
 from sentry.utils.safe import get_path, set_path
-from sentry.utils.strings import unescape_string
 
-from .actions import Action, FlagAction, VarAction
-from .exceptions import InvalidEnhancerConfig
-from .matchers import (
-    CalleeMatch,
-    CallerMatch,
-    ExceptionFieldMatch,
-    FrameMatch,
-    Match,
-    create_match_frame,
-)
+from .matchers import create_match_frame
+from .parser import parse_enhancements
+from .rules import Rule
 
 DATADOG_KEY = "save_event.stacktrace"
 logger = logging.getLogger(__name__)
@@ -45,52 +34,6 @@ logger = logging.getLogger(__name__)
 # NOTE: The 1_000 here is pretty arbitrary. Our builtin base enhancements have about ~300 rules,
 # So this leaves quite a bit of headroom for custom enhancement rules as well.
 RUST_CACHE = RustCache(1_000)
-
-# Grammar is defined in EBNF syntax.
-enhancements_grammar = Grammar(
-    r"""
-
-enhancements = line*
-
-line = _ (comment / rule / empty) newline?
-
-rule = _ matchers actions
-
-
-matchers         = caller_matcher? frame_matcher+ callee_matcher?
-frame_matcher    = _ negation? matcher_type sep argument
-matcher_type     = ident / quoted_ident
-caller_matcher   = _ "[" _ frame_matcher _ "]" _ "|"
-callee_matcher   = _ "|" _ "[" _ frame_matcher _ "]"
-
-actions          = action+
-action           = flag_action / var_action
-var_action       = _ var_name _ "=" _ ident
-var_name         = "max-frames" / "min-frames" / "invert-stacktrace" / "category"
-flag_action      = _ range? flag flag_action_name
-flag_action_name = "group" / "app" / "prefix" / "sentinel"
-flag             = "+" / "-"
-range            = "^" / "v"
-
-ident            = ~r"[a-zA-Z0-9_\.-]+"
-quoted_ident     = ~r"\"([a-zA-Z0-9_\.:-]+)\""
-
-comment          = ~r"#[^\r\n]*"
-
-argument         = quoted / unquoted
-quoted           = ~r'"([^"\\]*(?:\\.[^"\\]*)*)"'
-unquoted         = ~r"\S+"
-
-sep      = ":"
-space    = " "
-empty    = ""
-negation = "!"
-newline  = ~r"[\r\n]"
-_        = space*
-
-"""
-)
-
 
 VERSIONS = [2]
 LATEST_VERSION = VERSIONS[-1]
@@ -577,16 +520,7 @@ class Enhancements:
     def from_config_string(self, s, bases=None, id=None) -> Enhancements:
         rust_enhancements = parse_rust_enhancements("config_string", s)
 
-        try:
-            tree = enhancements_grammar.parse(s)
-            rules = EnhancementsVisitor().visit(tree)
-        except ParseError as e:
-            context = e.text[e.pos : e.pos + 33]
-            if len(context) == 33:
-                context = context[:-1] + "..."
-            raise InvalidEnhancerConfig(
-                f'Invalid syntax near "{context}" (line {e.line()}, column {e.column()})'
-            )
+        rules = parse_enhancements(s)
 
         return Enhancements(
             rules,
@@ -594,179 +528,6 @@ class Enhancements:
             id=id,
             rust_enhancements=rust_enhancements,
         )
-
-
-class Rule:
-    def __init__(self, matchers, actions):
-        self.matchers = matchers
-
-        self._exception_matchers = []
-        self._other_matchers = []
-        for matcher in matchers:
-            if isinstance(matcher, ExceptionFieldMatch):
-                self._exception_matchers.append(matcher)
-            else:
-                self._other_matchers.append(matcher)
-
-        self.actions = actions
-        self._is_updater = any(action.is_updater for action in actions)
-        self._is_modifier = any(action.is_modifier for action in actions)
-
-    @property
-    def matcher_description(self):
-        rv = " ".join(x.description for x in self.matchers)
-        for action in self.actions:
-            rv = f"{rv} {action}"
-        return rv
-
-    def _as_modifier_rule(self) -> Rule | None:
-        actions = [action for action in self.actions if action.is_modifier]
-        if actions:
-            return Rule(self.matchers, actions)
-        else:
-            return None
-
-    def _as_updater_rule(self) -> Rule | None:
-        actions = [action for action in self.actions if action.is_updater]
-        if actions:
-            return Rule(self.matchers, actions)
-        else:
-            return None
-
-    def as_dict(self):
-        matchers = {}
-        for matcher in self.matchers:
-            matchers[matcher.key] = matcher.pattern
-        return {"match": matchers, "actions": [str(x) for x in self.actions]}
-
-    def get_matching_frame_actions(
-        self,
-        match_frames: Sequence[dict[str, Any]],
-        exception_data: dict[str, Any],
-        in_memory_cache: dict[str, str],
-    ) -> list[tuple[int, Action]]:
-        """Given a frame returns all the matching actions based on this rule.
-        If the rule does not match `None` is returned.
-        """
-        if not self.matchers:
-            return []
-
-        # 1 - Check if exception matchers match
-        for m in self._exception_matchers:
-            if not m.matches_frame(match_frames, None, exception_data, in_memory_cache):
-                return []
-
-        rv = []
-
-        # 2 - Check if frame matchers match
-        for idx, _ in enumerate(match_frames):
-            if all(
-                m.matches_frame(match_frames, idx, exception_data, in_memory_cache)
-                for m in self._other_matchers
-            ):
-                for action in self.actions:
-                    rv.append((idx, action))
-
-        return rv
-
-    def _to_config_structure(self, version):
-        return [
-            [x._to_config_structure(version) for x in self.matchers],
-            [x._to_config_structure(version) for x in self.actions],
-        ]
-
-    @classmethod
-    def _from_config_structure(cls, tuple, version):
-        return Rule(
-            [Match._from_config_structure(x, version) for x in tuple[0]],
-            [Action._from_config_structure(x, version) for x in tuple[1]],
-        )
-
-
-class EnhancementsVisitor(NodeVisitor):
-    visit_comment = visit_empty = lambda *a: None
-    unwrapped_exceptions = (InvalidEnhancerConfig,)
-
-    def visit_enhancements(self, node, children) -> list[Rule]:
-        rules = []
-        for child in children:
-            if not isinstance(child, str) and child is not None:
-                rules.append(child)
-
-        return rules
-
-    def visit_line(self, node, children):
-        _, line, _ = children
-        comment_or_rule_or_empty = line[0]
-        if comment_or_rule_or_empty:
-            return comment_or_rule_or_empty
-
-    def visit_rule(self, node, children):
-        _, matcher, actions = children
-        return Rule(matcher, actions)
-
-    def visit_matchers(self, node, children):
-        caller_matcher, frame_matchers, callee_matcher = children
-        return caller_matcher + frame_matchers + callee_matcher
-
-    def visit_caller_matcher(self, node, children):
-        _, _, _, inner, _, _, _, _ = children
-        return CallerMatch(inner)
-
-    def visit_callee_matcher(self, node, children):
-        _, _, _, _, _, inner, _, _ = children
-        return CalleeMatch(inner)
-
-    def visit_frame_matcher(self, node, children):
-        _, negation, ty, _, argument = children
-        return FrameMatch.from_key(ty, argument, bool(negation))
-
-    def visit_matcher_type(self, node, children):
-        return node.text
-
-    def visit_argument(self, node, children):
-        return children[0]
-
-    def visit_action(self, node, children):
-        return children[0]
-
-    def visit_flag_action(self, node, children):
-        _, rng, flag, action_name = children
-        return FlagAction(action_name, flag, rng[0] if rng else None)
-
-    def visit_flag_action_name(self, node, children):
-        return node.text
-
-    def visit_var_action(self, node, children):
-        _, var_name, _, _, _, arg = children
-        return VarAction(var_name, arg)
-
-    def visit_var_name(self, node, children):
-        return node.text
-
-    def visit_flag(self, node, children):
-        return node.text == "+"
-
-    def visit_range(self, node, children):
-        if node.text == "^":
-            return "up"
-        return "down"
-
-    def visit_quoted(self, node, children):
-        return unescape_string(node.text[1:-1])
-
-    def visit_unquoted(self, node, children):
-        return node.text
-
-    def generic_visit(self, node, children):
-        return children
-
-    def visit_ident(self, node, children):
-        return node.text
-
-    def visit_quoted_ident(self, node, children):
-        # leading ! are used to indicate negation. make sure they don't appear.
-        return node.match.groups()[0].lstrip("!")
 
 
 def _update_frames_from_cached_values(
